refactor: separate regional briefing and weather sources
