+++ server/providers/regional/query.js
@@ -1,0 +1,11 @@
+import { requiredFiniteQueryNumber } from '../common/query.js';
+
+function validRegionalPoint(params) {
+  const latitude = requiredFiniteQueryNumber(params, 'latitude');
+  const longitude = requiredFiniteQueryNumber(params, 'longitude');
+  if (!Number.isFinite(latitude) || !Number.isFinite(longitude)) return null;
+  if (latitude < -90 || latitude > 90 || longitude < -180 || longitude > 180) return null;
+  return { latitude, longitude };
+}
+
+export { validRegionalPoint };
